# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PitchTZ Client API
  version: 0.2.0
  summary: Public and authenticated player API for the PitchTZ mobile and client-web apps.
  description: |
    This contract intentionally excludes admin, owner, internal provider-webhook, and payout APIs.
    Operations marked `x-pitchtz-status: live` are deployed today. Operations marked `planned`
    are the agreed mobile contract and will be implemented without changing their request shape.
    All money values are integer Tanzanian shillings and all timestamps are RFC 3339 UTC values.
servers:
  - url: https://pitchtz-production.up.railway.app/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
paths:
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Lock slots and create a pending booking
      operationId: createBooking
      x-pitchtz-status: planned
      security: [{ playerBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Pending booking created with a 10-minute payment lock.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "409":
          description: One or more requested slots are no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    playerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Player access JWT issued by OTP or OAuth. Admin JWTs are a different audience and are rejected.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A new UUID per user action. Reuse the same value when retrying the same action.
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: INVALID_INPUT }
            message: { type: string }
...
    Booking:
      type: object
      required: [id, code, pitch_id, starts_at, ends_at, status, pitch_fee_tzs, service_fee_tzs, total_tzs, payment_shares]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PITCH-7284 }
        pitch_id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        status: { type: string, enum: [pending, confirmed, part_paid, completed, cancelled, disputed] }
        repeat_rule: { type: string, enum: [once, weekly4, monthly, 3months] }
        pitch_fee_tzs: { type: integer, format: int64 }
        service_fee_tzs: { type: integer, format: int64 }
        total_tzs: { type: integer, format: int64 }
        lock_expires_at: { type: string, format: date-time }
        payment_shares: { type: array, items: { $ref: "#/components/schemas/PaymentShare" } }
    PaymentShare:
      type: object
      required: [id, amount_tzs, kind, status]
      properties:
        id: { type: string, format: uuid }
        amount_tzs: { type: integer, format: int64 }
        kind: { type: string, enum: [full, share, installment_1, installment_2, organizer_cover, tip_guards, tip_venue] }
        status: { type: string, enum: [unpaid, pending, paid, refunded, failed] }
        due_at: { type: string, format: date-time }
        qr_url: { type: string, format: uri }
        paid_at: { type: string, format: date-time }
...
    CreateBookingRequest:
      type: object
      required: [pitch_id, slots]
      properties:
        pitch_id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        slots:
          type: array
          minItems: 1
          maxItems: 12
          items:
            type: object
            required: [starts_at, ends_at]
            properties:
              starts_at: { type: string, format: date-time }
              ends_at: { type: string, format: date-time }
        extras:
          type: array
          items:
            type: object
            required: [extra_id, quantity]
            properties:
              extra_id: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1, maximum: 50 }
        repeat_rule: { type: string, enum: [once, weekly4, monthly, 3months], default: once }
        promo_code: { type: string }
...
    BookingResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties: { data: { $ref: "#/components/schemas/Booking" } }
...